[Res Explorer] Add filtering by file extension or xml tag
Adding model and api to be able to filter resources by file extension
(PNG, JPG, WEBP) or Xml root-tag.

Should be particularly useful when trying to search through multiple
types of Drawable resources.

This is just the logic so is not in use. Will integrate it with a UI on
follow-up CL.

Bug: 140158835
Test: ResourceExplorerListViewModelImplTest, FilterTypeOptionsModelTest
Change-Id: Ib126917a76dce7807617bcbb52d3b043ed1da732