[Gradle Dsl] Introduce IncludeDslElement.
Replace using GradleDslExpressionList with the newly created
includeDslElement. lists were used in order to maintain the order of
included modules but this resulted in:
When adding modules to the settingsModel, the dslWriter used to append
the newly created psiElement in the same line (ex: include ':app', 'mobile')
instead of creating a dedicated include instruction line for every
module which is the preffered syntax.

Fixes: 144280955
Change-Id: I1c1c7d64897ffad94a876c3a753b3a1927db42dc
Test: unit tests already existing.