Fix problem with Build attribution UI init
Fix a problem that info panel is not visible immediately after tab recreation.
The problem was in CardLayout used to show Build Attribution info panels.
Turned out that when there is only one element added to the layout ui elements validation is not triggered as expected
and thus panel remains invisible until next UI refresh.

Bug: N/A
Test: Manually verified
Change-Id: Iefa1654b6331b1173fc6841bb3c4923fa29f8717