Show IDE and Gradle time in sync perfgate
We are currently showing only total time in the initial and regular
benchmarks. This CL creates a new benchmark per scenario and splits
total time in the IDE and Gradle parts. As a summary this is what is
generated after this change:

- (New) 1 benchmark per scenario (currently we only have 2:
Average-Project-IDEA-sync Average-Project-Single-Variant-sync). Each of
these benchmarks will contain the following metrics:
  - Initial total time
  - Initial IDE time
  - Initial Gradle time
  - Regular total time
  - Regular IDE time
  - Regular Gradle time

- Initial and regular sync time benchmarks, used as summaries. They will
contain one metric per scenario.

Bug: 144360427
Test: Existing, manually ran bazel tests from cli.
Change-Id: I6833fcaaa1b4432667d3741a15d2891a7010b252