[Dsl] generalize compileOptions support
The BaseCompileOptionsModel/DslElement serves two functions: it
supports the compileOptions android block, but also models the
top-level (build file level) sourceCompatibility and
targetCompatibility properties.  This leads to some contortions in
this commit, because while the property inheritance is basically
correct, the Dsl surface is not the same in both cases.  Within the
android.compileOptions block, we have the usual setters, but at
top-level we only have properties, so we don't use the usual
inheritance in the JavaDsl pseudo-element.

For LanguageLevel properties, default to the JavaVersions.VERSION_X_Y
syntax if there are no other clues (particularly on element creation),
since this is the form that is accepted by the largest set of property
setters and variables -- I believe the whole set, but I have been
surprised before.

Bug: 141842964
Bug: 141970574
Bug: 142114586
Test: updated, strengthened, run for .kts CompileOptionsModel
Test: strengthened JavaModel
Change-Id: Ic4bd8cf686654a9d78c4872434fc35c3a9392690