Guard AnnotationPreviewElementFinder by read action
Accesses within the AnnotationPreviewElementFinder need to be within a
ReadAction when called from a thread that is not the dispatch one. This
will allow calling the method from anywhere.
Also moved the change listener to the background thread since it does
not need to be in the dispatch thread anymore.

Test: Existing tests pass
Bug: N/A
Change-Id: Iab00c32ed949c3e2539a1f64e55379df955fa4f6