Add a prefix to metric names to prevent overwrites
... when running more that one test on the same target.

When running tests from bazel, the output folder is shared amongst all
tests in the same target. When writting metrics to disk, they are
written to a file in the output folder named <metric_name>.json. In the
case of the performance tests we are getting the same metrics for
different projects, causing metrics from a previous test to be
overwritten (and thus lost).

This CL adds the scenario name as a prefix to the metric files so all
metrics are kept even when multiple tests are run in the same target.

This CL also simplifies some of the tests and renames metric labels to
be more homogeneous.

Test: Existing tests should work, manually confirmed the metrics are
kept.
Bug: 144360427

Change-Id: I38a37d90c1aa8d769387cd5a4903f742485e2ad2