Implement status bar and text in AndroidTestSuiteView
This CL implements the most basic functions in status bar and status text.
The UI needs to be polished and the view class needs to handle edge
cases correctly such as tool failures or no test cases found, etc. The
main purpose of this CL is to build a simplified UI with basic functions.

Fixes: 143298844
Test: AndroidTestSuiteViewTest. More test cases to be added later.
Change-Id: I6f968c336c3c22f205207f7b5e70df9afe864c6f