[Res Explorer] Add ConvertToWebP action to right click menu
Add the ConverToWebP action to the Resource Explorer right click menu.
WebP is the preferred format over JPG and PNG images.

This CL fixes an issue that would cause the right-click menu in the
ResourceDetailView (when clicking on a resource with multiple
configuration) to pick up the data from the previously selected asset,
caused by having the selection update after the popup menu was invoked.

Bug: 143494541
Test: manually
Change-Id: I02cdc5c3cd318a6ae9485a5cc1c0c6b03402eb27