Use simple application as warm up project
... when running sync perf tests.

We are seeing an increase of aroud 3x in time for the initial sync of
the first scenario run from intellij.android.sync-perf-tests. This CL
adds a new test case that will sync SIMPLE_APPLICATION multiple times
before running any other test to warm up the JVM and the gradle daemon
and have a more homogenous environment for all scenarios.

This adds ~40 seconds to every scenario, a total of 160 seconds.

Test: Existing, manually checked that initial sync times are more
consistent between different scenarios.
Bug: 144360427

Change-Id: I7d7abebcdfe53b791694eae303265815f86945ba