[Res Manager] Fix ShowFileInResourceManagerAction shorcut not working
The ShowServicesAction gets triggered before the Resource Manager
Action, and it consumes the keyboard event event if its associated tool
window is not visible or in use.

This makes it so the ShowServicesAction is put at the end of the queue
in the Event Dispatcher, so that the event can go through the Resource
Manager.

Bug: 140454601
Test: manually
Change-Id: I0e700e363587e6d1a4dad875f4dbe5ed6d7d4749
(cherry picked from commit 352ec1495b4398165f6d7257637c7957b5a57fd4)