[Dsl] generalize android.externalNativeBuild support
In terms of functionality this CL is a no-op: the model only currently
supports the path feature of ndkBuild and cmake, and there's nothing
particularly tricky about handling that.  Nevertheless, introduce an
appropriate superclass for CMakeDslElement and NdkBuildDslElement,
write the generalized handling and port most tests to
KotlinScript.  (Because the path property is of type File, there are a
couple of tests involving explicit constructors where we need to
handle the result of the constructor in the Dsl explicitly, which we
are not yet doing in KotlinScript).

Bug: 141842964
Bug: 141970574
Bug: 142114586
Test: existing ExternalNativeBuildModel, strengthened and run for .kts
Change-Id: I776b2b6dc24fcd557ef1a9c55317e6957192d590