[Kotlin Dsl Parser] Add support for File() constructor methodCalls.
As We can't distinguish between a constructor call and normal call
expressions using psi elements in KTS without resolving psi
references, add a special handling for File()
callExpressions at the parser level to create a constructor
dslMethodCall.
File() is the only callExpressions that we know of and we handle properly
for dsl models (i.e. GradleSettingsModel, CMakeModel, NDKBuildModel), so we
can assume it is safe for now using a special handling for constructor calls.

Bug:145395390
Fixes:144492499
Test: existing
Change-Id: I2f0e7430a6928f19b9b62fe607cc2b8db25ff443