[Dsl] generalize signingConfig support
The base properties all have the same name in Groovy and Kotlin, so
this is less of a change than the generalization for many other
properties.  Nevertheless, the existing tests for Groovy check that
method calls and assignments both work, and although I am not sure
that I would expect anyone to use the explicit setter in KotlinScript,
it is easy to support in this new implementation, so do so.

Write a getPasswordModelForProperty method in GradlePropertyModel by
analogy with getFileModelForProperty.  Strengthen the existing tests
to assert the contents of files with changes applied, and fix the
existing assertions to reflect the default method (Groovy) /
assignment (Kotlin) syntax.

Bug: 141842964
Bug: 141970574
Bug: 142114586
Test: existing SigningConfigModel tests, strengthened.
Change-Id: I31f03dc6933cc019ec3ebd3fc5343b0be03ff063