Split some code from Preview.kt to other files
Done in order to keep more clear when converting PreviewEditor
to a PreviewRepresentation. Only the PreviewEditor is left
in the Preview file. All the shared constants are moved to
Utils. ComposePreviewManager interface is split into a
separate file. ComposePreviewEditorNotificationAdapter and
ForceCompileAndRefreshAction are in their own separate files.
The rest of the code is in ComposeFileEditorProvider file.
Some minor visibility changes are also made.

Bug: 143067434
Test: all tests should pass
Change-Id: I630da3bfbec75709c4f5afda31744975a2792956