[Wizard Templates] Fix resource loading on release build
When running on "proper" release build
the resources are packed inside
"lib jar files" and loaded by a special
"plugin classloader". To find the resources
we need to supply a class that was loaded
by the right class loader.
When loaded this way, the resources need
to have a full path (ie start with "/").
This CL improves on Change-Id Ie35f51e420dce07b51f55ca79026e77eb21b370a

Bug: 135397531
Test: TemplateTests + manual from a release build
Change-Id: I93194f02a4b8f0b1ef08b139f4f167d5e1668943