[Dsl,.kts] support packagingOptionsModel set properties
Now that we have model support for distinguishing set properties from
list ones, we can support and test properly the packagingOptions
properties excludes, merges and pickFirsts.

One test is still currently disabled for KotlinScript because of an
unrelated bug, and some of the tests are not as strong as they could
be because of a different bug related to modelling the
packagingOptions operators.

Fixes: 143193213
Test: PackagingOptionsModel tests run for .kts
Change-Id: Ic422382359b13f649a908b2a260339c171ff5220