[Res Explorer] Add ui for the new type filters
Add the TypeFilter options to the Filters pop-up menu. On the top-right
corner of the Resource Explorer.

Enabled under new flag.

TypeFilters using the same display name are grouped together into one
option in the ui. There's also a 'Clear Filters' option to clear (set
to false) all filters for the current Resource Type in the Resource
Explorer.

Added some more filters, including filters for ConstraintLayout and
MotionLayout.

Note that these filters are only active on an instance basis. Meaning
that if the Resource Manager/Picker are closed, the options will be
reset to false.

Bug: 144788553
Bug: 140158835
Test: manually for ui, existing for logic
Change-Id: I7e5722360d6eabb36b4c4cfcf762dcded1e204d6