[Dsl] allow models to have a default
It is (arguably) the model's responsibility to decide what to do if
there is no underlying Dsl element.  Most of the time, the model will
disclaim all knowledge (say that the property corresponding to the
element is missing) but in some cases it is useful to present to the
caller the known-default value.

This commit simply implements the possibility for models to add
defaults on construction (and extracts the value if there is no
corresponding element).

Bug: 144679123
Test: existing
Change-Id: I217a509bd2a923682df5d9dfb660c71197cb6dcc