Fix warnings and reenable test
This CL fixes "Write-unsafe context" warnings
generated when running a WriteAction to perform an
APK copy. It also fixes the followup issue where
VFS is still syncing when shutting down, by forcing
the VFS to finish sync before progressing.

The test is reenabled by moving the test out of the
EXCLUDED group.

Fixes: 144739159
Bug: 144187508
Test: Verified all Write-unsafe and VFS warnings are fixed
Change-Id: I1fa628ba457c6703fbd97ced62be3b5a3471b5a7