Support ObjectTree class after package change
With commit (in intellij-community):
https://github.com/JetBrains/intellij-community/commit/11bc14ab8a8310172d566871ad72d483329e308b#diff-8a96aff72a3c38f54b32a8cb362e555a
(Note: We haven't integrated that change to Android Studio)

ObjectTree class is now in package com.intellij.openapi.util instead of
com.intellij.openapi.util.objecttree. HProf analyzer expected the
former and would throw an exception if not found.

Solution:
Support both package names for ObjectTree class. Also, if
AnalyzeDisposer will not be able to find any of the classes/fields,
it will report the exception as the part of the memory report.

Fixes: N/A
Test: Manual
Change-Id: I7dda31926b3d938612f755a859905e2572faf108