[Build Model] Support for setting module paths
This is a change to support the setting of the projectDir properties on
Projects in the Gradle settings file. This is required to remove the
old GradleBuildFile parser and related classes.

This change is not intended to be a permanent solution as it doesn't
support KTS and has to make some less than optimal compromises.

Test: unit tests added and updated
Bug: 136801101
Change-Id: Iad1c53f232e10c2522dafbba1fc3bd00e5282413