[Dsl,.kts] extend RepositoriesModel tests
Primarily, strengthen the existing tests for Groovy-based build syntax
to verify the contents of modified build files, in preparation for
some surgery required to support repository properties in
KotlinScript.  As an incidental by-product, we can add KotlinScript
versions for some tests where the syntax is sufficiently similar that
we already pass them.

Bug: 142114586
Bug: 144679123
Test: existing, strengthened, run for .kts
Change-Id: I202d14f6fb7542bedfba2f2dbe484c904a83d146