Make the default surface state configurable
The default DesignSurface state was being set to FULL in every scenario,
and it was responsibility of whoever instantiate the DesignSurface to change
the state afterwards if applicable. This CL changes this behavior to make
things more explicit: the default mode should be pass to the DesignSurface
when the instance is created, so the callers won't need additional logic
to handle an initial state change.

The split editor now uses the default surface state to determine the default
editor mode if "Prefer XML editor" is not explicitly selected.

Compose and Android Custom View Preview, as well as drawables, should
set the default state to SPLIT.

Test: Added unit test and updated existing
Bug: 144902711
Change-Id: I5d09d12b99cb4157cf3390da12b93e13c7e9f128