[Dsl,.kts] generalize support for Ndk
The NdkOptionsDslElement and Model currently only supports the
abiFilters property.  An unrelated bug in AGP's KotlinScript
support (specifically the support for cmake/ndkOptions
externalNativeBuild blocks) prevents us from running most of the
existing tests for KotlinScript, but we can at least verify that the
modified support for Ndk continues to function.

Bug: 141842964
Bug: 141970574
Test: one additional (almost trivial) KotlinScript test
Change-Id: I74ac97afb08d3b661c3c1905d8bb67baffdc7ac3