[Dsl] tabulate child PropertiesElement information
Clients can look up the description for a child properties block
corresponding to a String name using a new method in
GradleDslPropertiesElement.  This method has a default implementation,
suitable for most Dsl elements, which looks up the name in a map.  For
NamedDomainObject containers, this default method is overridden to
return a property description for the corresponding
NamedDomainElement (which by its nature does not have a static name).

This allows us to mostly replace the enormous block of code in
SharedParserUtils with a table lookup.  It will also allow us to
detect when people have misplaced blocks.

Bug: n/a
Test: existing
Change-Id: I0548403d13fae5f95eab7b2de1382ad14f1da700