Don't hold on to repositories in CachedValues used by light classes
ResourceRepositoryManager is free to drop its caches and create new
instances of repositories, which means old instances no longer increment
their modification counts and will not force CachedValues to be
recomputed. To fix this problem, we use an indirect ModifcationTracker
that returns the modification count of the "current" repository instance
instead of storing the (possibly stale) instance in the CachedValue.

This made me realize the API for tracking dependencies of CachedValues
used by light classes can be simplified, so I changed it to use a single
ModificationTracker which is all we need for now.

Fixes: 144585792
Test: LightClassesTests.kt
Change-Id: I65dd25e2afa20ac3a18a646954441c3a4e8f3d79