Guard array access when checking for binary manifest format
Currently, AndroidManifestIndex$Indexer checks the first two bytes
of a manifest file to see if it's in binary manifest format. We do
this without first making sure that the content has at least two
bytes, which causes ArrayIndexOutOfBoundsExceptions if the manifest
is empty.

This CL resolves the issue by updating the toManifestReader()
utility to explicitly check the content length before trying
to access the bytes.

Fixes: 145153123
Test: AndroidManifestIndexTest
Change-Id: I97e81d6f4ecc09690c01ef2a15a31842f5bb3909