123822220: Cyclic dependency for android:src and srcCompat
This CL removes the @Required attribute from the <bitmap>
tag in XML drawables.

That's because this tag isn't actually required in some
cases, for example when using app:srcCompat instead (which
will inject src on your behalf), and the presence of this
warning would lead developers to fight conflicting suggestions
("add src"! "No, use srcCompat instead!", as described in
issue 123822220.)

This means you no longer get a warning in the XML editor
if you don't specify either attribute; there's no simple
way to configure the DOM machinery to have conditional
required attribute handling. We could add a check to Android
Lint to warn about this missing attribute, but it's unlikely
that this leads to tricky to find bugs; the BitmapDrawable
class invokes #verifyRequiredAttributes at runtime and
produces a clear error message.

Fixes: 123822220
Test: Unit test included
Change-Id: I5ca07959269dd3bc8cdec5a0da902d3f4cbc9c65