[Windows] Fix template test date not matching
The "date()" used in the live template variables does not match on
different systems. According to JetBrains' template variables page,
"Without a parameter, the current date is returned in the default system
format." so these tests failed on Windows because the year format was
2019 instead of 19.

This CL changes the test to use the same DateFormatUtil to get the
expected value.

Bug: 141868211
Test: this
Change-Id: Ieccacb74bfb305c6e69660e05e45b1eb65cece93