Don't assume snapshot with null MergedManifestInfo is valid
If an error occurs while generating or parsing the merged manifest,
MergedManifestManager caches an "empty" snapshot and then never
recomputes due to a mistaken assumption that the module must be
disposed in MergedManifestManager#snapshotUpToDate.

This CL resolves the issue by explicitly checking to see if the
module is disposed in snapshotUpToDate(). If the module is still
active, then we know there was an error in the last computation
and that the snapshot isn't fresh.

Fixes: 144939434
Test: MergedManifestManagerTest

Change-Id: I24b368519a190982691a7d64f4de730837607cbe