[Dsl,.kts] support autocreation of map property entries
Convert manifestPlaceholders["foo"] and
testInstrumentationRunnerArguments["foo"] in Kotlin's gradleNameFor to
dotted notation.  (Only those two because they are the only two Dsl
map properties we currently support under KotlinScript.  This is not a
general solution, but allows an additional test to run to help in the
preparation for a more general solution.)

Bug: 143761530
Test: run a BuildTypeModel for KotlinScript
Change-Id: I74f447e7cbc3890197de18c58286def74a9ccec1