[Dsl] support dereferenced-map lvalues
The existing Groovy test case for directly setting entries in maps had
two defects.  Firstly, it involved invalid syntax: there is no setter
method for non-existent keys in maps.  Secondly, it only tested one of
the two syntaxes for directly setting values: property-like dotted
reference, but not array-like bracket reference.

Since bracket reference is legal, we should support it if we can, and
since it is the only way to refer literally to entries in maps in
Kotlinscript, we should aim to support it generically.  We already
support extracting values using this syntax, so we just have to adapt
our treatment of assignment lvalues to handle this case.  As a bonus,
that after doing that we can remove the low-level special-casing of
most specific strings in the Kotlin parser.

When we have an lvalue in Groovy or Kotlin, whose name after handling
qualifiers matches GradleNameElement.INDEX_PATTERN, we need to
sequentially extract properties corresponding to all but the last
index; the last index is then a reference to the element in the final
property that we need to add or modify.  Implement this, and then if
the resulting property is a map, add a (shadowing) property of the
appropriate name.  (Note that setting elements of lists remains
unsupported even after this commit.)

This use of GradleNameElement.INDEX_PATTERN is expedient but has the
(possibly-)minor defect of losing information about the underlying Psi
elements.  Each backend language should perhaps use its own parsing
machinery to extract the indices and corresponding elements, but this
solution works well enough to be tested, and to allow further
simplification of sub-block lookup.

Bug: 138277630
Bug: 143761530
Test: existing test strengthened
Change-Id: I7a2e85d83cae645ac13fb7879011732931848339