Simplify threading in Preview
This simplifies the threading model in the Preview in preparation for
the use of coroutines. With this CL we just have one refresh happening
in one executor that will automatically run them in sequence. When more
than one refresh is queuing, the oldest one will be discarded.

This also helps reducing the callback hell a little bit since now we are
happy to block the currently running render.

Test: Existing tests pass and new tests added on the base code
Bug: N/A
Change-Id: Ib4d23ebac368f72f859ec75f5769f575eacedc43