[Dsl,.kts] fix deletion inside significant-if-empty blocks
productFlavor (and some other) blocks should not be removed even if
they are empty, as their presence creates (implicitly in Groovy,
explicitly in KotlinScript) the corresponding ProductFlavor object.
In KotlinScript, we were checking the parent Dsl element (as opposed
to the current Dsl element) too soon, and keeping one level of block
hierarchy too many.

Bug: 138277630
Test: ProductFlavors.notRemoved test for KotlinScript
Change-Id: I01235d45fbaed4a0e6babfabeb8964fd98c4a3b3