Allow using both single and double quotes for project() properties.
At the dsl level, and when creating dsl elements for project() properties,
we use a canonical name that consists of using single quotes for argument.
However, and if we want to lookup the corresponding dsl element later at the
parsers level, we use the external property namefor lookup which won't work
if in the external name we use double quotes for the method argument.
The CL allows getting the canonical name at the dslNameConverters stage
to use for lookups.

Bug: n/a
Test: existing (the updated groovy testfile is to illustrate the fix
that was broken before this change)

Change-Id: I53501037ea0ad98750d502f14d82f38affad37ef