Update the key mapping for properties table.
Now that a PTableGroupItem can get focus in the NAME column the
keyboard navigation in the properties table is broken.
The keys mapped only works if the table has focus. We want the same
keys to also function if an editor is active and the editor doesn't
consume the key.

Bug: 145309578
Test: Existing tests should pass.
Change-Id: Ia3bc062cafa93796ec8b80a561a244910a005fba