Resolve plurals resource type to int with empty parameter list
In db-compiler's logic, the fall-back solution for plurals
resource without parameter is to resolve its type to int.
In studio-editor's however, we check the children of resolved
element instead of the original element, which cause
dbReferencesStringPluralResource to fail in some scenarios.
For example:
@plurals/orange(3) is resolved to String as
it has a child for parameters.
@plural/orage is resolved to int as it does not have any child
for parameters.
In studio's logic before this CL, however, we check the children
list of the resource's resolved element
<plurals name="orange">
  <item quantity="one">orange</item>
  <item quantity="other">oranges</item>
</plurals>

Fixes: 145628178
Test: DataBindingExprReferenceContributorTest
Change-Id: I8d4b621c999e362b752b2b8759fd46cb57a602b8