[Res Explorer] Fix leak in ImageCache Key
ImageCache should now use a combination of the 'DesignAsset' file path,
type and its resource name as key for the Cache.

This key should resolve the project being leaked when closed.

It should also be unique for different resources declared on the same
file (eg: theme attributes). As well as resources with the same name,
on the same file but of different type (eg: color and drawable references
in a values.xml file).

Bug: 142959061
Bug: 145698018
Test: ImageCacheTest for theme attributes bug regression
Change-Id: I084422b906d3820df540c370881cd3aedff14596