Filter nodes without node Ids from being added to the tree
This change allows us to create a root node to hold multiple threads when they are
selected. We do not want the leaf of each thread to trace back up to an empty root or an
invalid root as such we exclude it from the hierarchy

Also cleaning up some lint warnings.

Test: Updated
Bug: 145153662
Change-Id: Ic03bbab366fc90bf3fc4be71d3b7b34710f2cd5b