Kill client after socket accept timeout
Previously, if the client process encountered some problem that
prevented it from getting to the point where it could establish a
connection with the server, and did not quit of its own accord, it would
be left running. If enough instances (128) accumulate, it consumes all
available connections to the X server (not to mention the RAM).

This change ensures that clients get killed in this situation.

Change-Id: I4008216ede35bbcb7eb0736490deaf7da30fce45