Delegate a focus request on a non focusable text editor to parent
The table UI will handle a mouse click event by creating the
underlying cell editor and pass the focus request to the editor
based on the mouse click coordinates.

In the layout inspector the user will usually click on the embedded
PropertyTextField editor which will receive the focusRequest from
the table UI. The focus request would fail because the field it not
editable in the current version of the layout inspector.

In this CL, I will delegate the focusRequest to the first parent of
the text editor that is focusable.

Fixes: 145763044
Fixes: 145762947
Test: TODO: b/145828501
Change-Id: If7508695621d6dd4ecb7d87f5fd5361d38ac29df