Eliminate empty expanding attributes
The layout inspector is using a InspectorGroupPropertyItem to
display a property that has extra information. Such an item will
always have an expansion control.

In the simplest case the expanded item should show the source
location. Up to this CL when the source location could not be
found exactly the expanded item would contain ... nothing.

With this change an approximate source location will be added
to show either: "filename:?" or "unknown".
The user will be able to goto the link of "filename:?" which
will show the root tag of that xml file. But we are unable to
locate the exact attribute in that xml file.

Fixes: 145772244
Test: Added unit test
Change-Id: If90eb7371f06ae7a3e047edbc376573537df9c28