Change how file resources with invalid names are handled
All names that are valid Java identifiers will be put in the repository,
even if the names are not accepted by aapt. This should be less
confusing to users, allowing them to rename the resource from code etc.

At the same time, all file resources with invalid names are marked as
problematic with WolfTheProblemSolver service which is used by the
project view.

Bug: 144443057
Test: ResourceFolderRepositoryTest
Change-Id: I09fc1b15ba5a8309d70e4bf5cc6862587278fb01