Fix Kotlin NPE in CompileErrorsTest
An upstream commit caused the behavior around build error message
reporting to change. This, in turn, broke a test in our codebase which
was checking against the callstack to verify that a regex pattern was
found there.

In this CL, we now look at the underlying build error directly.

The upstream change can be seen here:
https://github.com/JetBrains/intellij-community/commit/4222f8e8dedaec621b78d68ef9d941bf620f71df#diff-09b0663d3d25f573718815c83282f5d5

Fixes: 145854390
Test: Updated to use new code; @Ignore removed
Change-Id: I4a03abc65738151b2b6c976540d06fa49822259a