[KTS Dsl Writer] Ensure we safely update dslLiteral.
If the dslLiteral value's psiElement is a typed element (ex: value as type1)
we don't consider the type decoration psi element for the literal
expression since it is useless for the value itself at the dsl level.
Make sure to take it into consideration again at the writer stage when
updating the literal.

Bug: n/a
Test: manually verified, utility will be shown in an upcoming CL.
Change-Id: I7ce22aa392c4595e0a28d4975df2f4534f38fcbc