[Dsl,.kts] generalize support for SourceDirectory elements
There is a particular challenge in supporting the srcDir function,
similar to a problem with proguardFile, in our current implementation:
it is a method which accepts a single argument, so the model should
not write newly-created elements to the end of the argument list of
srcDir.  At present, the Dsl model is unaware of this; to compensate,
change the order of calls in the test file so that we do not expose
the invalid syntax in these tests, and log a bug to document that we
do not get all cases right.

Also support setSrcDirs by analogy with setProguardFiles: it is a
special-case at the moment, but better understanding of what SET
functions do would be able to resolve this.

(As with SourceFiles, this Dsl syntax will be implemented for
KotlinScript in a pending change for AGP.)

Bug: 138277630
Bug: 141842964
Bug: 141970574
Test: existing tests strengthened, run for KotlinScript
Change-Id: Ibcfaff06c47608686dd88d03a9ee22d1fff6c23a