[KTS] Begin work on dsl reference resolution.
As the DslText is configured in a similar way to Groovy language syntax,
we never needed to do extra work to resolve dsl references and we would
just write the dslText to the Gradle script as it is. This does not work
for Kotlin and we need to, for each reference, to convert the text
reference to a Kotlin friendly syntax before writing.

This CL is a first step as it only focuses on resolving local variables
plus maps and literals extra properties.
For extra properties, the syntax can be quite tricky: If it happens that we
use array expressions to refer to a property, kotlin can't figure
out the property type (as Kotlin is a statically typed language) and we
would need to use explicit type cast in order to not break the project.
As we don't currently have an information about the target context type,
we are using the reference  property type for casting.

Bug:142454204
Test: Manually tested
Change-Id: Icfe3a829bbcca29565830330dd16047918a01e02