Fix CreateDefaultActivityTest Flakiness
When adding a new Activity to an existing module,
and pressing finish, there are two dialogs stacked,
the bottom one is the New Activity Wizard, and the
second one is a "Progress Dialog popup". When the
top dialog is dismissed, the second is also dismissed,
but for sometime is possible that a "dismissed" dialog
still has the Windows focus.

Bug: N/A
Test: CreateDefaultActivityTest (reproduces 1 in 10)
Change-Id: I6f604645e3be86d814acf851f16ae3597c81b22e