[Dsl,.kts] parsing method calls with named arguments
In order to re-use parts of the Dsl that were written assuming Groovy
syntax, parse first method calls where all arguments are named as
GradleDslExpressionMap objects, matching how Groovy application
statements would be handled.

This is not a long-term solution, but as well as helping to solve
round-tripping of apply(plugin="...") forms, it enables us to support
all the forms of writing artifact dependencies.

Bug: 141842964
Bug: 142539025
Test: existing tests pass
Change-Id: I829179734048117308b9fd0f2257e654040b2676