[Dsl,.kts] write maps as method calls if required
Some functionality (notably "apply plugin", but also dependency
notation) depends on being able to write out map-like data as a method
call with named arguments.  In the long term, we will probably remove
the pun of named arguments as maps, but in the shorter term creators
of maps with this requirement (at the moment, only
GradleBuildModelImpl) must specify this explicitly.

Bug: 141842964
Fixes: 142539025
Test: run some more ApplyPlugin tests for KotlinScript
Change-Id: Ic3d415d6b71deeab9e5aeec12f17a3999a6c1090