[Dsl,.kts] artifact dependency parsing tests
Some of the current tests for dependencies in Groovy-based build files
are not applicable to KotlinScript: Groovy accepts multiple
dependencies for a single configuration ("on one line").  KotlinScript
does not support this; its interface allows passing named arguments, a
map of expanded notation, or a string of compact notation.  Port the
remaining dependency parsing tests that apply to KotlinScript.  (Some
tests need minor language-specific adjustment because of the change in
syntax.)

Also strengthen one of the tests to verify more syntactic forms in
both Groovy and KotlinScript.

Bug: 138277630
Test: more, stronger tests now run for KotlinScript
Change-Id: If142773ae241ffff1ad4e92c61e1526298c7fb69