Suppress NlPreviewManager listeners if SplitEditor is in use
When we're using the SplitEditor, we shouldn't be displaying the Preview
tool window. Since that's the case, we should not create the listeners
for file/project open/close events in NlPreviewManager. In some scenarios,
these listeners would try to access the preview window, which was not set,
consequently resulting in a NullPointerException being thrown.

Fixes: 146228201
Test: Manual. NPE should not be thrown when the split editor flag is
      enabled.

Change-Id: I78216c235d5dfcebdb4c2a6d94cbccd6de3da50c