Make nav files from transitive deps merged manifest contributors
Currently, manifest merging in the IDE fails if a manifest includes
a <nav-graph> tag for an activity and the corresponding navigation
file includes a navigation graph from a dependency. This is because
we don't pass in navigation files from dependencies when invoking
ManifestMerger2, and so the merger can't resolve the id of the
included navigation graph.

This CL resolves the issue by updating MergedManifestContributors
to include navigation files from the module's transitive
dependencies.

Test: GradleMergedManifestContributorsTest
Bug: 143828478

Change-Id: Ica0e752f93919b085b07efa03eb67a092d9010da