Fix selection range max changing on resize min
The reason this happens is because the ClampedAxisComponentModel rounds
the range to the next interval set by the constructor of the model.
In the case we are using it, this adjust the selection range and we
don't want more than one element modifying this range. As such giving
the model its own range to round allows it to scale the AxisComponent
without modifing the selection range on update.

Test: added test to validate behavior
Fixes: 146214970
Change-Id: I443d589b69baa905ad12293e8523e8a39df045dd