Use created recycler view setup wizard
Recycler view has bit of a boiler plate codes. We allow
generation of alot of it but the options are hidden.
(new -> Fragment -> Fragment (List))

Also the option isn't always applicable. What if you do
not want a fragment since the reycler view you're creating
isn't the parent of the layout?

This CL adds 2 options that are directly accessible from
the right-click menu of the recycler view directly.

1) Add fragment as well as the item, adapter, etc (same as
Fragment (List))
2) Add adapter, list, item (which are the most common boiler
plate code)

(Creation CL of these templates are in ag/9942989)

Bug: 146472440
Test: Manually tested

Change-Id: I966edea612b34a23436a294ecf6ebfc68b9b96da