Strengthen the code to determine the current motion layout.
The bug b/143770265 has an attached project which has problems
rendering when a custom motion layout is opened in design without
an explicit rebuild.

The code was relying on the viewInfo which in this case could not
be determined by layoutlib. This CL will check the ViewHandler as
well to make the determination if this is a motion layout.

Also add a few null pointer checks.

Bug: 143770265
Test: N/A

Change-Id: I96b7ce51690dd65844525e31338be32f7916f9f0