Consolidate creation of OutputBuildAction.
With AGP 4.0, OutputBuildAction is deprecated and replaced with
getListingFile APIs. The OutPutBuildAction while running tasks
should not be used anymore.

This change consolidates the creation of OutputBuildAction into
one place, OutputBuildActionProvider. This will be the single
place to check AGP versions and decides whether a BuildAction should
be used or not.

This also fixes a problem in the previous code, where Gradle version
3.5 is checked to decide if BuildAction should be used.
This was problematic because the feature was added to AGP 3.0, AGP
prior to 3.0 can use Gradle 3.5, which will cause unsupported method
exception.
The new change checks for AGP version instead since AGP 3.0 requires
Gradle 4.1.

Bug: 146352662
Test: GradleTaskRunnerFactoryTest
Change-Id: If07a6a905ccec5d45ee0883fdad815816b20f249