[Dsl] remove a now-unused model creator
We can't yet remove the whole isMethod flag from the model: there are
still some direct uses of it.  However, all straightforward properties
are now handled by the model and the Dsl externalizers without
reference to the model's isMethod flag, so we no longer need a
creator method to set it.

Bug: 141970574
Test: existing (no change in functionality)
Change-Id: I262b1bac29be6d4cfd0cb5f85473a272211829fb