Avoid log spam in ManifestUtils#getActionLocation
Currently, the method assumes that the source files associated with
manifest merger actions can only be manifest files, but this isn't
true since, for example, we can have actions that produce intent
filters from deep links in navigation files. In such cases,
ManifestUtils#getActionLocation logs a warning. The method gets
called quite frequently to update the UI, so this results in a lot
of log spam.

This CL resolves the issue by adding special case handling for
source files that are not manifests to ManifestUtils#getActionLocation.
This CL also adds some comments to the method to make it more clear
what we're doing, as I found it to be non-obvious.

Bug: 143828478
Test: ManifestActionLocationTest
Change-Id: I9eab2e6ff551dea15ecd54e0575e5f0606e35f80