Fix flakiness of launchInspectorReturnsCorrectConnection
We were relying on executor to tell us whether listeners callbacks were queued. In the context of this test, we wanted to verify the executor queue is empty when we add a response with wrong commandId. However that introduced flake because executor is used for a bunch of other things (ex: launching an inspector connection). It would not be feasible to accomodate for them.

Therefore, reverting this CL back to its original form. That is to
assume executor executions in chronological order. We have two
connections waiting to be established c1 and c2, and we generate two
service responses r2 and r0 (r0 doesn't match with any connections). And
at the end, we check c2 is matched with r2, while c1 is left pending.

Reverted I64d72e69 and made some improvements on it.

Test: this is it.
Fixes: 144511139
Change-Id: I7b1d83f0f8499df9bb7429f4e2828ba8054e7f22