[Dsl,.kts] fileTreeDependency tests
The fileTree dependency model tests are oddly specific about the
fully-qualified names of properties -- "oddly" specific because there
doesn't seem to me to be any specific logic about when "fileTree"
itself was part of the name or not; I believe it was possibly an
artifact of whether a particular dependency was declared as a single
argument to a configuration name, or as one of multiple arguments.  (I
am not the only one exhibiting: there were existing comments querying
whether a particular string was the correct name.)

Since KotlinScript has yet different syntactic constraints about how
to declare dependencies, these tests have been weakened not to check
the name of the property.

Bug: 138277630
Bug: 142114586
Test: existing tests updated, run for KotlinScript
Change-Id: If450e6db2ee0318623bb87ca1f9abf0a07df2c14