Rename addModel to addAndRenderModel and add addModelWithoutRender
Give caller a chance to differ the render request.
This CL is for supporting sequential rendering in the following CLs.
Ideally we should trigger the modelChanged callback when NlModel is
added instead of NlModel is inflated. The callback for inflating
be ModelListener#modelDerivedDataChanged. Changing it is too risky
for now. Leave TODO with bug tracking number for it.

Bug: 143308628
Test: N/A (API change only)
Change-Id: I260b8c91ccb4be2047269ef167a74501746e1357