[KTS] Add support for resolving list references.
This CL adds support to resolving extra properties list elements in KTS
by converting the dslText to a text compatible with KTS syntax.
This CL also makes an improvement to allow using properties names as a
reference if they are being applied within the declaration context.
For example :
- If we have a map property defined in build.gradle
    val maps by extra(mapOf("key1" to "val1"))
  a reference to key1 within the same file will look like:
    maps["key1"]
  while a reference from a subModule file will look like:
    (parent.extra["maps"] as Map<*, *>)["key1"] as String

Nevertheless, in KTS we cannot use a property name if it is defined using array
access expression even if the reference is from the same context, so
this remains not supported in this CL.

Bug: 142454204
Test: Test added to verify list reference is correctly applied and read.
Change-Id: Ide552c7b9879b8a37552f9e41908c7b29638a15e