[PSD] detect obsolete "runtime" configuration
The Gradle `runtime` configuration (provided by the java plugin) is
deprecated; it had some of the same semantic problems as `compile`,
because it was intended to be a runtime-only dependency but
nevertheless was added to the compile-time classpath.

Detect dependencies with `runtime` (and related) configurations, and
offer quick fixes.  Ensure that all dependencies in Java modules, not
just libraries, are analyzed.

Bug: n/a
Test: added to ObsoleteConfigurationsTest
Change-Id: Ief83f64a86b6b592bdf62c8b8db5641c025ee23d