Speed up designer-perf-tests and fix deps
Reduce the number of samples taken for the compile operation since
most of it is I/O operations and hopefully the measure is still good
even with reduced samples.
If the deviation increases a lot, we'll revisit the CL.

Removed dependency on the core.tests and moved the plugin import into
the perf module.

Bug: N/A
Test: Updates existing tests
Change-Id: Ifbc2738b3ce2fd182a97deb042b51393b147dd51