Regenerate lint inspection registrations
When we add lint checks, a unit test is responsible for making sure that
each lint check is also registered as an IDE inspection, and if not, it
optionally updates the codebase to register it. However, over time a
number of checks have been tweaked after the registration was added, so
the two are not exactly synchronized.

This CL regenerates all this data, and places it in the canonical order.
This is in preparation for an upcoming CL which splits up the lint
support and separates the Android-specific inspections from the generic
inspections.

Test: Covered by existing tests.
Change-Id: If8c1d09d5c2eb9682a6bb2a1a65922c5e26a23b9