Change percentage calculation on Build Attribution charts
On time distribution charts for tasks determining build duration and plugins with such tasks
we present percentage of the time corresponding item takes.
Initially 100% was the total build time which often results in big percentage going to
'Gradle misc steps' containing configuration time, this data is not connected
to tasks and looks confusing for the user.
Change it so that 100% is tasks execution phase on these charts, not total build time.

Bug: 144393368
Test: updated
Change-Id: I165c6e46e2f178a02e9137ee23f5fd7ee0ec84d5