Don't add null module to BuildAction in MakeBeforeRunTaskProvider
Before this CL, MakeBeforeRunTaskProvider calls DynamicAppUtils#getBaseFeature
and puts that return value to the target build module list without checking
null. Because DynamicAppUtils#getBaseFeature returns null if a given
module is not a dynamic feature module, this causes an NPE when you run an
instrumentation test against a standard module (non-dynamic feature module).

Fixes: 146659758
Test: GradleTaskRunnerFactoryTest
Change-Id: I01ec2bf2c47b271d30eceaaebbe8f702081fc01c