Rework LibraryDependency names for local dependencies
so that (1) they are always derived from the artifact path, (2) the
path is relative to the project if possible and (2) does not contain
colons (':').

Also, even though it should never happen that two libraries with the
same name have different artifacts, fix the way we fix up their names so
that they don't become `Gradle: 2_Gradle: library_name`.

Bug: 141868211
Test: existing
Change-Id: Ic19a72d79ff3cedbf3e55d2c83bfd55df800100f