Fix data binding cache logic causing ConcurrentModificationException
In a previous CL, I wrote logic which I thought was thread-safe because
it was behind a lock, but I returned a collection outside of that lock
which was modified in-place within it.

Essentially, this CL changes the cache logic from this:

    val _groups = mutableSetOf()
    val groups {
      synchronized (lock) {
        _groups.clear()
        _groups.doStuff()
      }
      return _groups
    }

to:

**  var _groups = emptySet()
    val groups {
     synchronized (lock) {
**    _groups = mutableSetOf()
      _groups.doStuff()
     }
     return _groups
    }

This way, if a caller happened to be holding onto an older version of
the set of groups, it would be slightly stale, but they'd be allowed to
read it without issue.

Fixes: 146672451
Test: N/A - unfortunately, it's tricky to catch the threading issue, but
existing tests at leasy verify that behavior didn't break.

Change-Id: Iad08ec2a5552fbb85a4c6577e16d24b6b43b6f94
(cherry picked from commit 0bd329dc380292640dc0f5e4d3622c2ec18171b5)